+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project are documented in this file.
 - Prepare for github app integration with second binary(pr [#127])
 - add GitHub API integration with octocrate(pr [#128])
 - add AuthSource enum for authentication source and refactor(pr [#136])
+- add new_with_github_app function to create GitHub API client(pr [#138])
 
 ### Changed
 
@@ -288,6 +289,7 @@ All notable changes to this project are documented in this file.
 [#135]: https://github.com/jerusdp/ghdash/pull/135
 [#136]: https://github.com/jerusdp/ghdash/pull/136
 [#137]: https://github.com/jerusdp/ghdash/pull/137
+[#138]: https://github.com/jerusdp/ghdash/pull/138
 [Unreleased]: https://github.com/jerusdp/ghdash/compare/0.1.7...HEAD
 [0.1.7]: https://github.com/jerusdp/ghdash/compare/0.1.6...0.1.7
 [0.1.6]: https://github.com/jerusdp/ghdash/compare/0.1.5...0.1.6